+++ Modules/Core/ViewConfig.php
@@ -105,6 +105,7 @@ class ViewConfig extends ViewConfig_parent
             or $this->sCookieManagerType    === ManagerTypes::USERCENTRICS_MANUALLY
             or $this->sCookieManagerType    === ManagerTypes::CONSENTMANAGER
             or $this->sCookieManagerType    === ManagerTypes::COOKIEFIRST
+            or $this->sCookieManagerType    === ManagerTypes::COOKIEBOT
             or $this->sCookieManagerType    === ManagerTypes::EXTERNAL_SERVICE
         )
         {
@@ -152,6 +153,10 @@ class ViewConfig extends ViewConfig_parent
 
         if ($this->sCookieManagerType === ManagerTypes::COOKIEFIRST){
             return 'type="text/plain" data-cookiefirst-category="' . $sCookieId .'"';
+        }
+
+        if ($this->sCookieManagerType === ManagerTypes::COOKIEBOT){
+            return 'type="text/plain" data-cookieconsent="' . $sCookieId .'"';
         }
 
         return "";
@@ -207,6 +212,11 @@ class ViewConfig extends ViewConfig_parent
         ];
 
         return json_encode([$dataLayer], JSON_PRETTY_PRINT);
+    }
+
+    public function isDebugModeOn() :bool
+    {
+        return Registry::getConfig()->getConfigParam('d3_gtm_blEnableDebug');
     }
 
     public function isPromotionList($listId)